+++ src/mjlab/tasks/tracking/config/g1/__init__.py
@@ -1,8 +1,14 @@
 from mjlab.tasks.registry import register_mjlab_task
 from mjlab.tasks.tracking.rl import MotionTrackingOnPolicyRunner
 
-from .env_cfgs import unitree_g1_flat_tracking_env_cfg
-from .rl_cfg import unitree_g1_tracking_ppo_runner_cfg
+from .env_cfgs import (
+  unitree_g1_flat_tracking_env_cfg,
+  unitree_g1_table_tennis_tracking_env_cfg,
+)
+from .rl_cfg import (
+  unitree_g1_table_tennis_tracking_ppo_runner_cfg,
+  unitree_g1_tracking_ppo_runner_cfg,
+)
 
 register_mjlab_task(
   task_id="Mjlab-Tracking-Flat-Unitree-G1",
@@ -17,5 +23,13 @@ register_mjlab_task(
   env_cfg=unitree_g1_flat_tracking_env_cfg(has_state_estimation=False),
   play_env_cfg=unitree_g1_flat_tracking_env_cfg(has_state_estimation=False, play=True),
   rl_cfg=unitree_g1_tracking_ppo_runner_cfg(),
+  runner_cls=MotionTrackingOnPolicyRunner,
+)
+
+register_mjlab_task(
+  task_id="Mjlab-Tracking-TableTennis-Unitree-G1",
+  env_cfg=unitree_g1_table_tennis_tracking_env_cfg(),
+  play_env_cfg=unitree_g1_table_tennis_tracking_env_cfg(play=True),
+  rl_cfg=unitree_g1_table_tennis_tracking_ppo_runner_cfg(),
   runner_cls=MotionTrackingOnPolicyRunner,
 )